Backend/requirements.txt
# Django & REST Framework
Django==4.2.23
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Database
dj-database-url==3.0.1
psycopg2-binary==2.9.10

# Production
gunicorn==23.0.0
python-dotenv==1.0.1

# Utilities
sqlparse==0.5.3
asgiref==3.8.1
